Inventory Modifications
Modified all item functions related to inventory in order to make them be more widely used instead of just for the ammo calculations and such. Still need to implement adding onto a value such after opening the inventory. Also need to cleanup items in the array that reach 0 and then reset the ui to look how it did originally.