Implementation of AmmoInterface
This will allow it so anything that takes ammo can just be added into this interface in order to see if they can fire. So this will allow any type of gun pretty much to be added to this with just a few tweaks of visual effects and code changes that will have to do with that specific gun.